+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/EmptyTypeSectionRuleTest.java
@@ -1,0 +1,38 @@
+package org.sonar.plugins.delphi.pmd.rules;
+
+import static org.sonar.plugins.delphi.utils.conditions.RuleKeyAtLine.ruleKeyAtLine;
+
+import org.junit.Test;
+import org.sonar.plugins.delphi.utils.builders.DelphiTestUnitBuilder;
+
+public class EmptyTypeSectionRuleTest extends BasePmdRuleTest {
+  @Test
+  public void testRegularTypeSectionShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("type")
+            .appendDecl("  TClass = class(TObject)")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
+  }
+
+  @Test
+  public void testEmptyTypeSectionShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("type")
+            .appendDecl("  TClass = class(TObject)")
+            .appendDecl("  type")
+            .appendDecl("    // Empty nested type section")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues()
+        .hasSize(1)
+        .areExactly(1, ruleKeyAtLine("EmptyTypeSectionRule", builder.getOffsetDecl() + 3));
+  }
+}